move is_func_relevant to header
so we can test it without building entire PIN
