+++ FuncTracer.cpp
@@ -2,10 +2,8 @@
 #include "pin.H"
 #include <iostream>
 #include <fstream>
-#include <string>
-#include <set>
-#include <mutex>
 #include <unistd.h> // For getpid()
+#include "Functracer.hpp"
 
 // Global set and mutex to track logged functions
 static std::set<std::string> logged_functions;
@@ -31,21 +29,6 @@ VOID log_function_call(const char* img_name, const char* func_name)
     LOG(ss.str());
 }
 
-// Determine if function name is relevant to us and if it will be logged
-bool func_is_relevant(const std::string_view &func_name)
-{
-    static const std::set<std::string_view> blacklist = {
-        "main", "_init", "_start", ".plt.got"
-    };
-    if (blacklist.contains(func_name))
-        return false;
-
-    // Ignore PLT functions and internal functions (usually prefixed with __)
-    if (func_name.ends_with("@plt") || func_name.starts_with("__"))
-        return false;
-
-    return true;
-}
 
 // Pin calls this function for every image loaded into the process's address space.
 // An image is either an executable or a shared library.
